chore(ui): move Checkbox to ui folder
